# Documentazione/apidocs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /events/{id}:
    get:
      tags:
      - Events
      summary: Dettaglio evento
      description: "Include sede, organizer, relatori, tag, posti disponibili e valutazione\
        \ media."
      operationId: getEventById
      parameters:
      - name: id
        in: path
        description: ID evento
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Evento trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
        "404":
          description: Evento non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
    put:
      tags:
      - Events
      summary: Modifica un evento
      description: ORGANIZER (solo i propri eventi) o ADMIN. Stesse validazioni della
        creazione.
      operationId: updateEvent
      parameters:
      - name: id
        in: path
        description: ID evento
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRequestDTO"
        required: true
      responses:
        "200":
          description: Evento aggiornato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
        "400":
          description: Dati non validi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
        "403":
          description: Non sei il creatore dell'evento
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
        "404":
          description: Evento non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
    delete:
      tags:
      - Events
      summary: Elimina un evento
      description: ORGANIZER (solo i propri eventi) o ADMIN.
      operationId: deleteEvent
      parameters:
      - name: id
        in: path
        description: ID evento
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Evento eliminato
        "403":
          description: Non sei il creatore dell'evento
        "404":
          description: Evento non trovato
...
  /feedbacks:
    post:
      tags:
      - Feedbacks
      summary: Lascia un feedback
      description: |
        Permette all'utente autenticato di recensire un evento.
        Vincoli: evento concluso, biglietto attivo presente, nessun feedback duplicato.
      operationId: leaveFeedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequestDTO"
        required: true
      responses:
        "201":
          description: Feedback registrato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeedbackResponseDTO"
        "400":
          description: Voto non nel range 1-5
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeedbackResponseDTO"
        "404":
          description: Evento non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeedbackResponseDTO"
        "409":
          description: "Evento non concluso, nessun biglietto valido o feedback già\
            \ presente"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeedbackResponseDTO"
  /events:
    get:
      tags:
      - Events
      summary: Lista eventi pubblici
      description: |
        Restituisce tutti gli eventi con posti disponibili calcolati.
        Filtri opzionali combinabili: date (ISO-8601), tagId, venueId.
        Esempio: /events?date=2025-06-01T00:00:00&tagId=3
      operationId: getAllEvents
      parameters:
      - name: date
        in: query
        description: Filtra eventi con startDate > questa data (es. 2025-06-01T00:00:00)
        required: false
        schema:
          type: string
          format: date-time
      - name: tagId
        in: query
        description: Filtra per ID tag
        required: false
        schema:
          type: integer
          format: int32
      - name: venueId
        in: query
        description: Filtra per ID sede
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Lista eventi
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventResponseDTO"
        "400":
          description: Formato data non valido
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventResponseDTO"
    post:
      tags:
      - Events
      summary: Crea un nuovo evento
      description: ORGANIZER o ADMIN. La data di fine deve essere successiva alla
        data di inizio.
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRequestDTO"
        required: true
      responses:
        "201":
          description: Evento creato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
        "400":
          description: Dati non validi o date incoerenti
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
        "404":
          description: "Venue, tag o speaker non trovati"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
        "403":
          description: Accesso negato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventResponseDTO"
...
  /events/{id}/participants:
    get:
      tags:
      - Events
      summary: Lista partecipanti di un evento
      description: ORGANIZER (solo i propri eventi) o ADMIN. Restituisce i biglietti
        attivi.
      operationId: getEventParticipants
      parameters:
      - name: id
        in: path
        description: ID evento
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Lista biglietti attivi
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketResponseDTO"
        "403":
          description: Non sei il creatore dell'evento
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketResponseDTO"
        "404":
          description: Evento non trovato
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketResponseDTO"
  /events/{eventId}/rating:
    get:
      tags:
      - Feedbacks
      summary: Valutazione media di un evento
      description: Pubblica. Restituisce la media dei voti (1-5).
      operationId: getEventRating
      parameters:
      - name: eventId
        in: path
        description: ID evento
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Media calcolata (null se nessun feedback presente)
          content:
            '*/*':
              schema:
                type: number
                format: double
        "404":
          description: Evento non trovato
          content:
            '*/*':
              schema:
                type: number
                format: double
  /events/{eventId}/feedbacks:
    get:
      tags:
      - Feedbacks
      summary: Feedback di un evento
      description: Pubblica. Restituisce tutti i feedback lasciati per l'evento.
      operationId: getEventFeedbacks
      parameters:
      - name: eventId
        in: path
        description: ID evento
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Lista feedback
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedbackResponseDTO"
        "404":
          description: Evento non trovato
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedbackResponseDTO"
  /events/my:
    get:
      tags:
      - Events
      summary: I miei eventi
      description: Lista degli eventi creati dall'organizer autenticato. ORGANIZER
        o ADMIN.
      operationId: getMyEvents
      responses:
        "200":
          description: Lista eventi dell'organizer
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventResponseDTO"
        "403":
          description: Accesso negato
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventResponseDTO"
...
  /admin/feedbacks/{id}:
    delete:
      tags:
      - Feedbacks
      summary: Elimina un feedback
      description: Solo ADMIN. Moderazione contenuti inappropriati.
      operationId: deleteFeedback
      parameters:
      - name: id
        in: path
        description: ID feedback
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Feedback eliminato
        "404":
          description: Feedback non trovato
        "403":
          description: Accesso negato
components:
  schemas:
...
    EventRequestDTO:
      type: object
      description: Dati per la creazione o modifica di un evento
      properties:
        title:
          type: string
          description: Titolo dell'evento
          example: Workshop Spring Boot
          minLength: 1
        description:
          type: string
          description: Descrizione dell'evento
        startDate:
          type: string
          format: date-time
          description: Data e ora di inizio
          example: 2026-06-01T09:00:00
        endDate:
          type: string
          format: date-time
          description: Data e ora di fine
          example: 2026-06-01T18:00:00
        standardPrice:
          type: number
          description: Prezzo biglietto standard
          example: 49.99
          minimum: 0.0
        vipPrice:
          type: number
          description: Prezzo biglietto VIP
          example: 99.99
          minimum: 0.0
        venueId:
          type: integer
          format: int32
          description: ID della sede
          example: 1
        tagIds:
          type: array
          description: Lista ID dei tag
          example:
          - 1
          - 2
          items:
            type: integer
            format: int32
        speakerIds:
          type: array
          description: Lista ID dei relatori
          example:
          - 1
          - 2
          items:
            type: integer
            format: int32
      required:
      - endDate
      - standardPrice
      - startDate
      - title
      - venueId
      - vipPrice
    EventResponseDTO:
      type: object
      description: Dati di un evento
      properties:
        id:
          type: integer
          format: int32
          description: ID evento
          example: 1
        title:
          type: string
          description: Titolo
          example: Workshop Spring Boot
        description:
          type: string
          description: Descrizione
        startDate:
          type: string
          format: date-time
          description: Data e ora di inizio
          example: 2026-06-01T09:00:00
        endDate:
          type: string
          format: date-time
          description: Data e ora di fine
          example: 2026-06-01T18:00:00
        standardPrice:
          type: number
          description: Prezzo standard
          example: 49.99
        vipPrice:
          type: number
          description: Prezzo VIP
          example: 99.99
        availableSeats:
          type: integer
          format: int32
          description: Posti ancora disponibili
          example: 42
        venue:
          $ref: "#/components/schemas/VenueResponseDTO"
          description: Sede dell'evento
        organizer:
          $ref: "#/components/schemas/UserResponseDTO"
          description: Organizzatore dell'evento
        tags:
          type: array
          description: Tag dell'evento
          items:
            $ref: "#/components/schemas/TagResponseDTO"
        speakers:
          type: array
          description: Relatori dell'evento
          items:
            $ref: "#/components/schemas/SpeakerResponseDTO"
    SpeakerResponseDTO:
      type: object
      description: Dati di un relatore
      properties:
        id:
          type: integer
          format: int32
          description: ID relatore
          example: 1
        firstName:
          type: string
          description: Nome
          example: Mario
        lastName:
          type: string
          description: Cognome
          example: Rossi
        bio:
          type: string
          description: Biografia
          example: Senior developer con 10 anni di esperienza
        photoUrl:
          type: string
          description: URL foto
          example: https://example.com/foto.jpg
    TagResponseDTO:
      type: object
      description: Dati di un tag
      properties:
        id:
          type: integer
          format: int32
          description: ID tag
          example: 1
        name:
          type: string
          description: Nome del tag
          example: Java
    UserResponseDTO:
      type: object
      description: Dati pubblici di un utente
      properties:
        id:
          type: integer
          format: int32
          description: ID utente
          example: 1
        email:
          type: string
          description: Email utente
          example: mario@example.com
        role:
          type: string
          description: Ruolo utente
          enum:
          - USER
          - ORGANIZER
          - ADMIN
          example: USER
        status:
          type: string
          description: Stato utente
          enum:
          - ACTIVE
          - BANNED
          example: ACTIVE
    VenueResponseDTO:
      type: object
      description: Dati di una sede
      properties:
        id:
          type: integer
          format: int32
          description: ID sede
          example: 1
        name:
          type: string
          description: Nome della sede
          example: Centro Congressi Roma
        address:
          type: string
          description: Indirizzo
          example: "Via Roma 1, Roma"
        capacity:
          type: integer
          format: int32
          description: Capienza massima
          example: 200
...
    FeedbackRequestDTO:
      type: object
      description: Dati per lasciare un feedback su un evento
      properties:
        rating:
          type: integer
          format: int32
          description: Voto da 1 a 5
          example: 4
          maximum: 5
          minimum: 1
        comment:
          type: string
          description: Commento opzionale
          example: Evento molto interessante!
        eventId:
          type: integer
          format: int32
          description: ID dell'evento
          example: 1
      required:
      - eventId
      - rating
    FeedbackResponseDTO:
      type: object
      description: Dati di un feedback
      properties:
        id:
          type: integer
          format: int32
          description: ID feedback
          example: 1
        rating:
          type: integer
          format: int32
          description: Voto
          example: 4
        comment:
          type: string
          description: Commento
          example: Evento molto interessante!
        user:
          $ref: "#/components/schemas/UserResponseDTO"
          description: Utente che ha lasciato il feedback
        eventId:
          type: integer
          format: int32
          description: ID evento recensito
          example: 1
        eventTitle:
          type: string
          description: Titolo evento recensito
          example: Workshop Spring Boot
...
    TicketResponseDTO:
      type: object
      description: Dati di un biglietto prenotato
      properties:
        id:
          type: integer
          format: int32
          description: ID biglietto
          example: 1
        type:
          type: string
          description: Tipo di biglietto
          enum:
          - STANDARD
          - VIP
          example: STANDARD
        price:
          type: number
          description: Prezzo pagato
          example: 49.99
        status:
          type: string
          description: Stato del biglietto
          enum:
          - ACTIVE
          - CANCELLED
          example: ACTIVE
        eventId:
          type: integer
          format: int32
          description: ID evento
          example: 1
        eventTitle:
          type: string
          description: Titolo evento
          example: Workshop Spring Boot
...